dependency updates, refactoring, es6 (#7)
- authn: replace async.detect with local fn
- aliases: replace async with Promise.all
- dep(async): removed
- dep(ldapjs): update from 1.0.2 to 2.3.3
- noop: remove useless returns
- es6: more arrow functions
- style: inline use of `plugin`
- ci: switch travis -> GHA
- ci: enable codeql
- test: use RFC example.com for test domain
- test: refactored some to improve error messages
- test: add fixtures for setting up slapd on macosx and linux
- doc(README): update with GHA badge

## Test Failure

There's a test failure thats a bit difficult to track down. It manifests
when the entire test suite is run and doesn't when a single test file is
run. I'm not sure if the concurrency / race condition is within this
module or ldapjs but I've no more time to track it down.